+++ test/main.js
@@ -1,4 +1,4 @@
-import './fixtures/v0.4.6/input'
+import './fixtures/v0.4.7/input'
 
 document.querySelector('#app').innerHTML = `
   <h1>Hey Vite 👋</h1>
